+++ src/apis/kakaoLoginApi.ts
@@ -1,4 +1,4 @@
-import { UserType } from '@/pages/kakao/containers/types/kakaoLoginType';
+import { UserType } from '@/pages/register/containers/types/kakaoLoginType';
 import { client } from '@/apis/client';
 
 export const kakaoAuthCodeApi = (authCode: string) => {
