+++ src/pages/index.ts
@@ -1,7 +1,7 @@
 import { MainPage } from './main/page';
-import { KakaoRegisterPage } from './kakao/page';
-import { GeneralRegisterPage } from './general/page';
-import KakaoRedirect from './kakao/containers/KakaoRedirect';
+import { KakaoRegisterPage } from './register/page';
+import { GeneralRegisterPage } from './register/register-sort/page';
+import KakaoRedirect from './register/auth-callback/page';
 import { IntroPage } from './intro/page';
 import { IntroEditPage } from './intro/IntroEdit/page';
 import { AuditPage } from './audit/page';
@@ -32,7 +32,7 @@ import { LostDetailPage } from './lost-article/detail/page';
 import { LostPatchPage } from './lost-article/patch/page';
 import PersonalDataPage from './personal-data/page';
 import { Layout } from './layout/headerLayout';
-import { KakaoRegisterRedirectPage } from './kakao/redirect/page';
+import { KakaoRegisterRedirectPage } from './register/register-redirect/page';
 import MyPage from './mypage/page';
 import DataPage from './data/page';
 import DataDetailPage from './data/[id]/page';
