+++ src/pages/register/containers/RegisterTextSection.tsx
