+++ src/pages/register/containers/const/data.ts
