+++ src/pages/register/containers/utils/moveToRedirectUrl.ts
