+++ src/pages/register/register-sort/components/TOSContent.tsx
