+++ src/pages/register/register-sort/containers/TOSSection.tsx
